(WIP) Functional modification to use LowBandwidthTrajectoryFollower
Modified and added code to use LowBandwidthTrajectoryFollower with this RTDE implementation (instead of ros_control, based on implementation of ur_modern_driver).

Not completely tested yet, to be refactored. This is some kind of first 'alpha' version.

Note, that in this version the ros_control write method is ripped out. Trajectory execution is done using the trajectory follower.
Dashboard currently commented out.